requirements.txt
Pillow==9.3.0
streamlit==1.15.1
